feat: add support for rendering trades with triangles and improve Y bounds calculation
- Add TriangleRenderer to visualize trades as triangles colored by side (buy/sell)
- Fix Y bounds calculation to only exclude metrics that are exclusively additional data
- Implement render loop state machine for better control flow
- Add async data fetching with proper WASM handling using spawn_local
- Fix GPU bounds readback using staging buffers for actual min/max values
- Ensure all data groups are marked as active for multi-metric rendering
- Fix vertex buffer size mismatches by using minimum of X and Y buffer sizes
- Add preset and symbol configuration to DataStore
- Refactor LineGraph to ChartEngine for better naming clarity

The TriangleRenderer automatically finds data groups with "price" and "side" metrics
and renders triangles at trade positions. The Y bounds calculation now correctly
includes primary metrics even if they're also used as additional data columns.

+++ crates/data-manager/src/binary_parser.rs
@@ -115,8 +115,8 @@ pub fn binary_search_timestamp(data: &[u8], target: u32) -> Option<usize> {
 pub struct DataResponseHeader {
     pub symbol: String,
     pub columns: Vec<String>,
-    pub start_time: u64,
-    pub end_time: u64,
+    pub start_time: u32,
+    pub end_time: u32,
     pub row_count: usize,
 }
